Moved Application metrics to their respective model package
