+++ service/device/call/kura/src/main/java/org/eclipse/kapua/service/device/call/kura/model/command/CommandMetrics.java
@@ -10,7 +10,7 @@
  * Contributors:
  *      Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.service.device.call.kura.app;
+package org.eclipse.kapua.service.device.call.kura.model.command;
 
 import org.eclipse.kapua.service.device.call.message.DeviceMetrics;
 import org.eclipse.kapua.service.device.call.message.app.DeviceAppMetrics;
