+++ translator/kapua/kura/src/main/java/org/eclipse/kapua/translator/kapua/kura/TranslatorAppPackageKapuaKura.java
@@ -14,7 +14,7 @@
 package org.eclipse.kapua.translator.kapua.kura;
 
 import org.eclipse.kapua.model.id.KapuaId;
-import org.eclipse.kapua.service.device.call.kura.app.PackageMetrics;
+import org.eclipse.kapua.service.device.call.kura.model.deploy.PackageMetrics;
 import org.eclipse.kapua.service.device.call.message.kura.app.request.KuraRequestChannel;
 import org.eclipse.kapua.service.device.call.message.kura.app.request.KuraRequestMessage;
 import org.eclipse.kapua.service.device.call.message.kura.app.request.KuraRequestPayload;
