+++ translator/kapua/kura/src/main/java/org/eclipse/kapua/translator/kura/kapua/TranslatorAppNotifyKuraKapua.java
@@ -17,11 +17,11 @@ import org.eclipse.kapua.locator.KapuaLocator;
 import org.eclipse.kapua.model.id.KapuaIdFactory;
 import org.eclipse.kapua.service.account.Account;
 import org.eclipse.kapua.service.account.AccountService;
-import org.eclipse.kapua.service.device.call.kura.app.AssetMetrics;
-import org.eclipse.kapua.service.device.call.kura.app.BundleMetrics;
-import org.eclipse.kapua.service.device.call.kura.app.CommandMetrics;
-import org.eclipse.kapua.service.device.call.kura.app.ConfigurationMetrics;
-import org.eclipse.kapua.service.device.call.kura.app.PackageMetrics;
+import org.eclipse.kapua.service.device.call.kura.model.asset.AssetMetrics;
+import org.eclipse.kapua.service.device.call.kura.model.bundle.BundleMetrics;
+import org.eclipse.kapua.service.device.call.kura.model.command.CommandMetrics;
+import org.eclipse.kapua.service.device.call.kura.model.configuration.ConfigurationMetrics;
+import org.eclipse.kapua.service.device.call.kura.model.deploy.PackageMetrics;
 import org.eclipse.kapua.service.device.call.message.kura.app.notification.KuraNotifyChannel;
 import org.eclipse.kapua.service.device.call.message.kura.app.notification.KuraNotifyMessage;
 import org.eclipse.kapua.service.device.call.message.kura.app.notification.KuraNotifyPayload;
